projects/happy-kids-viet/requirements.txt
moviepy
Pillow
numpy
requests
python-dotenv
